Reorganize tests and RPC calls
Puts them into corresponding traits and directories
based on functionality and version (v16/v17)
