+++ bitcoind-rpc-test/src/test/scala/org/bitcoins/rpc/BitcoindRpcTestConfig.scala
@@ -1,0 +1,10 @@
+package org.bitcoins.rpc
+
+import scala.concurrent.duration.{DurationInt, FiniteDuration}
+import scala.util.Properties
+
+object BitcoindRpcTestConfig {
+
+  val DEFAULT_TIMEOUT: FiniteDuration =
+    if (Properties.isMac) 5.minutes else 1.minute
+}
