+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/v17/BitcoindV17RpcClient.scala
@@ -1,0 +1,68 @@
+package org.bitcoins.rpc.client.v17
+
+import akka.actor.ActorSystem
+import org.bitcoins.core.crypto.ECPrivateKey
+import org.bitcoins.core.protocol.BitcoinAddress
+import org.bitcoins.core.protocol.transaction.Transaction
+import org.bitcoins.rpc.client.common.{
+  BitcoindRpcClient,
+  BitcoindVersion,
+  RpcOpts
+}
+import org.bitcoins.rpc.config.BitcoindInstance
+import org.bitcoins.rpc.jsonmodels.{AddressInfoResult, SignRawTransactionResult}
+import org.bitcoins.rpc.serializers.JsonSerializers._
+import play.api.libs.json.{JsString, Json}
+
+import scala.concurrent.Future
+import scala.util.Try
+
+/**
+  * This class is compatible with version 0.17 of Bitcoin Core.
+  */
+class BitcoindV17RpcClient(override val instance: BitcoindInstance)(
+    implicit
+    actorSystem: ActorSystem)
+    extends BitcoindRpcClient(instance)
+    with V17LabelRpc {
+
+  override val version: BitcoindVersion = BitcoindVersion.V17
+  require(version == instance.getVersion, "bitcoind version must be 0.17")
+
+  def getAddressInfo(address: BitcoinAddress): Future[AddressInfoResult] = {
+    bitcoindCall[AddressInfoResult]("getaddressinfo",
+                                    List(JsString(address.value)))
+  }
+
+  def signRawTransactionWithWallet(
+      transaction: Transaction,
+      utxoDeps: Vector[RpcOpts.SignRawTransactionOutputParameter] = Vector.empty,
+      sigHash: String = ""
+  ): Future[SignRawTransactionResult] =
+    bitcoindCall[SignRawTransactionResult]("signrawtransactionwithwallet",
+                                           List(JsString(transaction.hex),
+                                                Json.toJson(utxoDeps),
+                                                Json.toJson(sigHash)))
+
+  def signRawTransactionWithKey(
+      transaction: Transaction,
+      keys: Vector[ECPrivateKey],
+      utxoDeps: Vector[RpcOpts.SignRawTransactionOutputParameter] = Vector.empty,
+      sigHash: String = ""
+  ): Future[SignRawTransactionResult] =
+    bitcoindCall[SignRawTransactionResult]("signrawtransactionwithkey",
+                                           List(JsString(transaction.hex),
+                                                Json.toJson(keys),
+                                                Json.toJson(utxoDeps),
+                                                Json.toJson(sigHash)))
+
+}
+
+object BitcoindV17RpcClient {
+
+  def fromUnknownVersion(rpcClient: BitcoindRpcClient)(
+      implicit actorSystem: ActorSystem): Try[BitcoindV17RpcClient] =
+    Try {
+      new BitcoindV17RpcClient(rpcClient.instance)
+    }
+}
